refactor: address pr feedback
